requirements.txt
numpy
matplotlib
tensorflow
tensorflow_hub
python-opencv2
pillow